Refactored Exciton into abstract class + ExcitonTB (not finished)
